Write file in cwd and move to final destination
Also fewer checkpoints.

This is in response to the network problems we've been having.  The
done "locally" (assuming like our condor configuration where the cwd is
a local drive) then a few attemptes are made to use the unix mv command
to move to the remove location.  mv makes all the necessary checks.

it might not have been a great idea to try 5 times if mv is hanging;
might be making the jobs take too long